+++ AlpacaService/Infrastructure/Alpaca/Contracts/IAlpacaTradingService.cs
@@ -12,18 +12,18 @@ namespace BN.PROJECT.AlpacaService
 
         Task<IAsset> GetAssetBySymbolAsync(string symbol);
 
-        Task<List<AlpacaOrder>> GetAllOrdersAsync(UserSettingsModel userSettings, OrderStatusFilter orderStatusFilter);
+        Task<List<AlpacaOrder>> GetAllOrdersAsync(OrderStatusFilter orderStatusFilter);
 
-        Task<AlpacaOrder> GetOrderByIdAsync(UserSettingsModel userSettings, string orderId);
+        Task<AlpacaOrder> GetOrderByIdAsync(string orderId);
 
-        Task<bool> CancelOrderByIdAsync(UserSettingsModel userSettings, Guid orderId);
+        Task<bool> CancelOrderByIdAsync(Guid orderId);
 
-        Task<AlpacaOrder> CreateOrderAsync(UserSettingsModel userSettings, string symbol, OrderQuantity qty, OrderSide side, OrderType orderType, TimeInForce timeInForce);
+        Task<AlpacaOrder> CreateOrderAsync(string symbol, OrderQuantity qty, OrderSide side, OrderType orderType, TimeInForce timeInForce);
 
-        Task<List<AlpacaPosition>> GetAllOpenPositions(UserSettingsModel userSettings);
+        Task<List<AlpacaPosition>> GetAllOpenPositions();
 
-        Task<AlpacaPosition> GetPositionsBySymbol(UserSettingsModel userSettings, string symbol);
+        Task<AlpacaPosition> GetPositionsBySymbol(string symbol);
 
-        Task<AlpacaOrder> ClosePositionOrder(UserSettingsModel userSettings, string symbol);
+        Task<AlpacaOrder> ClosePositionOrder(string symbol);
     }
 }