+++ AlpacaService/Infrastructure/Alpaca/Implementation/AlpacaTradingService.cs
@@ -72,9 +72,9 @@ namespace BN.PROJECT.AlpacaService
             return await tradingClient.GetAssetAsync(symbol);
         }
 
-        public async Task<List<AlpacaOrder>> GetAllOrdersAsync(UserSettingsModel userSettings, OrderStatusFilter orderStatusFilter)
+        public async Task<List<AlpacaOrder>> GetAllOrdersAsync(OrderStatusFilter orderStatusFilter)
         {
-            var tradingClient = _alpacaClient.GetPrivateTradingClient(userSettings);
+            var tradingClient = _alpacaClient.GetCommonTradingClient();
             var req = new ListOrdersRequest
             {
                 OrderStatusFilter = orderStatusFilter
@@ -84,45 +84,45 @@ namespace BN.PROJECT.AlpacaService
             return orders.Select(order => order.ToAlpacaOrder()).ToList();
         }
 
-        public async Task<AlpacaOrder> GetOrderByIdAsync(UserSettingsModel userSettings, string orderId)
+        public async Task<AlpacaOrder> GetOrderByIdAsync(string orderId)
         {
-            var tradingClient = _alpacaClient.GetPrivateTradingClient(userSettings);
+            var tradingClient = _alpacaClient.GetCommonTradingClient();
             var order = await tradingClient.GetOrderAsync(orderId);
             return order.ToAlpacaOrder();
         }
 
-        public async Task<bool> CancelOrderByIdAsync(UserSettingsModel userSettings, Guid orderId)
+        public async Task<bool> CancelOrderByIdAsync(Guid orderId)
         {
-            var tradingClient = _alpacaClient.GetPrivateTradingClient(userSettings);
+            var tradingClient = _alpacaClient.GetCommonTradingClient();
             return await tradingClient.CancelOrderAsync(orderId);
         }
 
-        public async Task<AlpacaOrder> CreateOrderAsync(UserSettingsModel userSettings, string symbol, OrderQuantity qty, OrderSide side, OrderType orderType, TimeInForce timeInForce)
+        public async Task<AlpacaOrder> CreateOrderAsync(string symbol, OrderQuantity qty, OrderSide side, OrderType orderType, TimeInForce timeInForce)
         {
-            var tradingClient = _alpacaClient.GetPrivateTradingClient(userSettings);
+            var tradingClient = _alpacaClient.GetCommonTradingClient();
             var req = new NewOrderRequest(symbol, qty, side, orderType, timeInForce);
             var order = await tradingClient.PostOrderAsync(req);
             var alpacaOrder = order.ToAlpacaOrder();
             return alpacaOrder;
         }
 
-        public async Task<List<AlpacaPosition>> GetAllOpenPositions(UserSettingsModel userSettings)
+        public async Task<List<AlpacaPosition>> GetAllOpenPositions()
         {
-            var tradingClient = _alpacaClient.GetPrivateTradingClient(userSettings);
+            var tradingClient = _alpacaClient.GetCommonTradingClient();
             var positions = await tradingClient.ListPositionsAsync();
             return positions.Select(p => p.ToAlpacaPosition()).ToList();
         }
 
-        public async Task<AlpacaPosition> GetPositionsBySymbol(UserSettingsModel userSettings, string symbol)
+        public async Task<AlpacaPosition> GetPositionsBySymbol(string symbol)
         {
-            var tradingClient = _alpacaClient.GetPrivateTradingClient(userSettings);
+            var tradingClient = _alpacaClient.GetCommonTradingClient();
             var pos = await tradingClient.GetPositionAsync(symbol);
             return pos.ToAlpacaPosition();
         }
 
-        public async Task<AlpacaOrder> ClosePositionOrder(UserSettingsModel userSettings, string symbol)
+        public async Task<AlpacaOrder> ClosePositionOrder(string symbol)
         {
-            var tradingClient = _alpacaClient.GetPrivateTradingClient(userSettings);
+            var tradingClient = _alpacaClient.GetCommonTradingClient();
             var deletePositionRequest = new DeletePositionRequest(symbol);
             var order = await tradingClient.DeletePositionAsync(deletePositionRequest);
             return order.ToAlpacaOrder();