+++ AlpacaService/Infrastructure/ServiceClients/Contracts/IFinAIServiceClient.cs
@@ -5,5 +5,7 @@
         Task<string?> TestOptimizationAsync();
         Task<string?> CreateDataframeAsync(StrategySettingsModel testSettings);
         Task<string> StartOptimizerAsync(StrategySettingsModel testSettings);
+        Task<string?> StartAlpacaPaperTradingAsync(string strategyName);
+        Task<string?> StopAlpacaPaperTradingAsync();
     }
 }