+++ AlpacaService/Migrations/20260907130219_BarPerSeconds.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using BN.PROJECT.AlpacaService;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
@@ -11,9 +12,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace BN.PROJECT.AlpacaService.Migrations
 {
     [DbContext(typeof(AlpacaDbContext))]
-    partial class AlpacaDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260907130219_BarPerSeconds")]
+    partial class BarPerSeconds
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -88,6 +91,50 @@ namespace BN.PROJECT.AlpacaService.Migrations
                     b.ToTable("Bars");
                 });
 
+            modelBuilder.Entity("BN.PROJECT.AlpacaService.AlpacaBarSeconds", b =>
+                {
+                    b.Property<long>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("bigint");
+
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
+
+                    b.Property<decimal>("C")
+                        .HasColumnType("numeric");
+
+                    b.Property<decimal>("H")
+                        .HasColumnType("numeric");
+
+                    b.Property<decimal>("L")
+                        .HasColumnType("numeric");
+
+                    b.Property<decimal>("N")
+                        .HasColumnType("numeric(20,0)");
+
+                    b.Property<decimal>("O")
+                        .HasColumnType("numeric");
+
+                    b.Property<string>("Symbol")
+                        .IsRequired()
+                        .HasColumnType("text");
+
+                    b.Property<DateTime>("T")
+                        .HasColumnType("timestamp with time zone");
+
+                    b.Property<decimal>("V")
+                        .HasColumnType("numeric");
+
+                    b.Property<decimal>("Vw")
+                        .HasColumnType("numeric");
+
+                    b.HasKey("Id");
+
+                    b.HasIndex("Symbol", "T")
+                        .IsUnique();
+
+                    b.ToTable("BarsPerSeconds");
+                });
+
             modelBuilder.Entity("BN.PROJECT.AlpacaService.AlpacaCalendar", b =>
                 {
                     b.Property<long>("Id")
@@ -114,6 +161,36 @@ namespace BN.PROJECT.AlpacaService.Migrations
                     b.HasKey("Id");
 
                     b.ToTable("Calendars");
+                });
+
+            modelBuilder.Entity("BN.PROJECT.AlpacaService.AlpacaExecutionModel", b =>
+                {
+                    b.Property<Guid>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uuid");
+
+                    b.Property<string>("Assets")
+                        .IsRequired()
+                        .HasColumnType("text");
+
+                    b.Property<DateTime>("EndDate")
+                        .HasColumnType("timestamp with time zone");
+
+                    b.Property<DateTime>("StartDate")
+                        .HasColumnType("timestamp with time zone");
+
+                    b.Property<Guid>("StrategyId")
+                        .HasColumnType("uuid");
+
+                    b.Property<int>("StrategyType")
+                        .HasColumnType("integer");
+
+                    b.Property<Guid>("UserId")
+                        .HasColumnType("uuid");
+
+                    b.HasKey("Id");
+
+                    b.ToTable("Executions");
                 });
 
             modelBuilder.Entity("BN.PROJECT.AlpacaService.AlpacaOrder", b =>